[nix] switch to building a flake
for our next trick, we will need an improvement from nixpkgs-unstable
for rustPlatform.buildRustPackage. instead of trying to backport it
ourselves we can simply use the version from the unstable branch for the
time being. (on top of that, it'd be really nice to expose particular
revisions of nixpkgs and other dependencies to users.)