[nix] support git dependencies
this required that we pull in bleeding-edge nixpkgs for
an update to rustPlatform.buildRustPackage, which added a
cargoLock.allowBuiltinFetchGit option. this escape hatch of an option
allows you to avoid needing to tell nix the hash of the dependency.

as part of this, i moved the rust build out into its own flake output,
so i can debug it directly more easily. i also moved oxalica/rust-
overlay to the flake inputs, although i need to either do nix registry
shenanigans to get it to automatically update (like nixpkgs).