image name changes, tagging, and minor refactoring
the iso9660 isn't really maintained, and one of its deps (nom v5) won't
build on a future version of rust. remove that dependency; we were only
using it to read version.txt out of the image.

since we are using flakes now, we can actually get the nixpkgs commit
(and all of our other deps' commits) from the regenerated lock file
after the nix build. instead of dealing with modifying serde_json::Value
we set up some derive structs to keep the code simpler.

we also move `crate::build::Output` into `crate::nix::Metadata`, because
all the metadata we care about from the build is the output file (see
next paragraph) and information about the nix build itself (nixpkgs
version, store hash of the resulting ISO, etc).

in preparation of making it so we can run `nix build` directly or via
docker, we give our functions an open file to write the ISO to instead
of a tempdir to play in.

we also change the image name to be only the package name and the nix
store hash of the resulting image. this should prevent duplicate image
uploads and unnecessary deploys once we start using dropkick in CI.
ideally we would append the ext4 filesystem within nix so the hash
contains that process as well.

to keep the metadata about all our dependencies, we add these as tags on
the registered AMI.