replace rustPlatform.buildRustPackage with crane
fixes #9.

while `allowBuiltinFetchGit` is mostly fine it doesn't quite do the
right thing when you have a git dependency with workspace deps; the
package directory is copied to the vendor tree, but the workspace
Cargo.toml with all the workspace deps is not.

probably should have done this in the first place, but we are now using
a rust-building derivation that is meant for general usage of building
cargo projects in nix, and not for building cargo projects in _nixpkgs_.