pass --package instead of --bin to cargo
i added `--bin ${binName}` originally to work around a build issue in
progenitor (which i still need to file a bug for!) that stopped us from
building a workspace crate with a client and a server. only the server
is necessary, so i ended up picking `--bin` over `--package` somewhat
arbitrarily.

crane uses a two-step build process so that rebuilds that change the
rust code, but not any dependencies, don't need to rebuild all the
dependencies. it does this by replacing the source tree with what are
essentially no-op lib and bin targets to build and save the dependency
artifacts. this had the effect of removing the bin we were targeting
with `--bin`, causing compilation to fail.

i didn't notice this at first, because -- and for reasons i don't fully
understand -- the service we are focusing on deploying at present did
not trigger this bug, but a naïve, minimum-dependency service did!
i think it's specific to the more complex service listing a specific
`[[bin]]` target in its Cargo.toml, which left the bin target present.
on the naïve crate, no bin target by the name we were expecting existed
anymore.

anywho fewer knowable things worked before this commit than do now, and
i have issues to investigate and file.