cdk: allow accessing instance tags via IMDS
Hardcoding resource names can be problematic, especially if you want to
run two of the same CDK stack in the same account. Equally problematic
is the inability for software on an EC2 instance to know how to find a
resource in its stack.

With this change the DropkickInstance construct grants the instance
the ability to read its tags. The construct and downstream consumers
can set tags on the instance to tell software running on the instance
what resources to use. We will use this as part of durable certificate
storage for Caddy.